requirements.txt
# FastAPI y servidor
fastapi
uvicorn[standard]

# LangChain y LangGraph
langchain
langchain-openai
langchain-groq
langgraph

# Bases de datos
sqlalchemy
mysql-connector-python
psycopg2-binary
pgvector
redis

# Análisis de datos y visualización
pandas
numpy
plotly
sentence-transformers
torch

# Utilidades
python-dotenv
pydantic
pydantic-settings
sqlparse
tenacity

# Testing
pytest
pytest-asyncio
pytest-cov
httpx

sqlparse

# Testing
pytest
pytest-asyncio
pytest-cov
httpx
